Переписал Update, Добавил ответы сервера для OrderServlet, Отрефакторил OrderService, Исправил баг с XmlMapper, Добавил Flyway, написал SessionFilter

+++ src/main/java/com/inobitec/orderxml/servlet/OrderServlet.java
@@ -5,8 +5,8 @@ import com.inobitec.orderxml.model.Message;
 import com.inobitec.orderxml.model.Order;
 import com.inobitec.orderxml.service.OrderServiceImpl;
 import lombok.RequiredArgsConstructor;
+import org.springframework.transaction.annotation.Transactional;
 
-import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
@@ -20,44 +20,72 @@ import java.io.PrintWriter;
 @RequiredArgsConstructor
 public class OrderServlet extends HttpServlet {
 
+    private static final long serialVersionUID = 1L;
+
     private final OrderServiceImpl orderServiceImpl;
 
-    @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        response.setContentType("text/xml");
-        response.setCharacterEncoding("UTF-8");
-        Integer id = Integer.valueOf(request.getParameter("id"));
-        Order order = orderServiceImpl.getOrderById(id);
-        XmlMapper xmlMapper = new XmlMapper();
-        PrintWriter out = response.getWriter();
-        String xmlOrder = xmlMapper.writeValueAsString(order);
-        out.print(xmlOrder);
-    }
+    private static final XmlMapper xmlMapper = new XmlMapper();
+
+    private static final String XML_TEXT_TYPE = "text/xml";
+
+    private static final String UTF_8 = "UTF-8";
+
+    private static final String ID_XML_PROPERTY = "id";
 
     @Override
+    @Transactional
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        XmlMapper xmlMapper = new XmlMapper();
+        response.setContentType(XML_TEXT_TYPE);
+        response.setCharacterEncoding(UTF_8);
         InputStream in = request.getInputStream();
+        PrintWriter out = response.getWriter();
         Message message = xmlMapper.readValue(in, Message.class);
+        Order orderFromRequest = message.getBody().getOrder();
         if ("create".equals(message.getCommand())) {
-            orderServiceImpl.saveOrder(message.getBody().getOrder());
+            orderServiceImpl.addOrder(orderFromRequest);
+            response.setStatus(HttpServletResponse.SC_CREATED);
         }
-    }
 
-    @Override
-    protected void doPut(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        XmlMapper xmlMapper = new XmlMapper();
-        InputStream in = request.getInputStream();
-        Message message = xmlMapper.readValue(in, Message.class);
+        if ("read".equals(message.getCommand())) {
+            Integer id = Integer.valueOf(request.getParameter(ID_XML_PROPERTY));
+            Order order = orderServiceImpl.getOrderById(id);
+            if (order == null) {
+                response.setStatus(HttpServletResponse.SC_NOT_FOUND);
+                out.print(response.getStatus());
+                return;
+            }
+            String xmlOrder = xmlMapper.writeValueAsString(order);
+            out.print(xmlOrder);
+            response.setStatus(HttpServletResponse.SC_OK);
+            return;
+        }
+
+        if ("delete".equals(message.getCommand())) {
+            Integer id = Integer.valueOf(request.getParameter(ID_XML_PROPERTY));
+            Order order = orderServiceImpl.getOrderById(id);
+            if (order == null) {
+                response.setStatus(HttpServletResponse.SC_NOT_FOUND);
+                out.print(response.getStatus());
+                return;
+            }
+            orderServiceImpl.deleteOrderById(id);
+            response.setStatus(HttpServletResponse.SC_OK);
+            return;
+        }
+
         if ("update".equals(message.getCommand())) {
-            Integer orderId = Integer.valueOf(request.getParameter("id"));
-            orderServiceImpl.updateOrder(orderId, message.getBody().getOrder());
+            Integer id = Integer.valueOf(request.getParameter(ID_XML_PROPERTY));
+            Order order = orderServiceImpl.getOrderById(id);
+            if (order == null) {
+                response.setStatus(HttpServletResponse.SC_NOT_FOUND);
+                out.print(response.getStatus());
+                return;
+            }
+            orderServiceImpl.updateOrder(id, orderFromRequest);
+            response.setStatus(HttpServletResponse.SC_ACCEPTED);
+        } else {
+            response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+            out.print(response.getStatus());
         }
-    }
-
-    @Override
-    protected void doDelete(HttpServletRequest request, HttpServletResponse response){
-        Integer id = Integer.valueOf(request.getParameter("id"));
-        orderServiceImpl.deleteOrderById(id);
     }
 }
